refactor: 마크다운 렌더러를 유틸리티로 분리
- src/utils/renderMarkdown.tsx 생성
- renderInline, renderMarkdown 함수 분리
- LegalSheet에서 import로 변경
- 재사용성 및 테스트 용이성 향상
