+++ src/components/LegalSheet/index.tsx
@@ -3,8 +3,9 @@
  * React Island (client:load)
  */
 
-import React, { useState, useEffect, useCallback, useRef } from 'react';
+import { useState, useEffect, useCallback, useRef } from 'react';
 import { PRIVACY_POLICY, TERMS_OF_SERVICE } from './content';
+import { renderMarkdown } from '../../utils/renderMarkdown';
 
 type SheetType = 'privacy' | 'terms' | null;
 
@@ -113,61 +114,6 @@ export function LegalSheet({ initialOpen = null }: LegalSheetProps) {
       window.removeEventListener('open-terms-of-service', handleOpenTerms);
     };
   }, [openSheetByType]);
-
-  // 인라인 볼드 처리 함수
-  const renderInline = (text: string): React.ReactNode => {
-    if (!text.includes('**')) return text;
-    const parts = text.split(/\*\*(.*?)\*\*/g);
-    return (
-      <>
-        {parts.map((part, j) =>
-          j % 2 === 1 ? <strong key={j}>{part}</strong> : part
-        )}
-      </>
-    );
-  };
-
-  // 마크다운을 HTML로 변환 (간단한 변환)
-  const renderMarkdown = (md: string) => {
-    return md
-      .split('\n')
-      .map((line, i) => {
-        // 헤딩
-        if (line.startsWith('## ')) {
-          return <h2 key={i} className="legal-h2">{line.slice(3)}</h2>;
-        }
-        if (line.startsWith('# ')) {
-          return <h1 key={i} className="legal-h1">{line.slice(2)}</h1>;
-        }
-        // 리스트 (볼드보다 먼저 체크 - 리스트 내 볼드 지원)
-        if (line.startsWith('- ')) {
-          return <p key={i} className="legal-li">• {renderInline(line.slice(2))}</p>;
-        }
-        if (/^\d+\. /.test(line)) {
-          const match = line.match(/^(\d+)\./);
-          const num = match ? match[1] : '';
-          return <p key={i} className="legal-li">{num}. {renderInline(line.replace(/^\d+\. /, ''))}</p>;
-        }
-        // 볼드 (리스트가 아닌 경우)
-        if (line.includes('**')) {
-          return <p key={i} className="legal-p">{renderInline(line)}</p>;
-        }
-        // 구분선
-        if (line === '---') {
-          return <hr key={i} className="legal-hr" />;
-        }
-        // 이탤릭
-        if (line.startsWith('*') && line.endsWith('*') && !line.startsWith('**')) {
-          return <p key={i} className="legal-p legal-italic">{line.slice(1, -1)}</p>;
-        }
-        // 빈 줄
-        if (line.trim() === '') {
-          return <br key={i} />;
-        }
-        // 일반 텍스트
-        return <p key={i} className="legal-p">{line}</p>;
-      });
-  };
 
   if (!openSheet) return null;
